+++ src/Exception/InvalidConfigException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace TomPHP\ConfigServiceProvider\Exception;
+
+class InvalidConfigException extends RuntimeException
+{
+}