+++ src/FileReader.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace TomPHP\ConfigServiceProvider;
+
+interface FileReader
+{
+    /**
+     * @param string $filename
+     *
+     * @return array
+     */
+    public function read($filename);
+}