+++ tests/FileLocatorTest.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace tests\TomPHP\ConfigServiceProvider;
+
+use PHPUnit_Framework_TestCase;
+use TomPHP\ConfigServiceProvider\FileLocator;
+
+final class FileLocatorTest extends PHPUnit_Framework_TestCase
+{
+    use TestFileCreator;
+
+    /**
+     * @var FileLocator
+     */
+    private $locator;
+
+    protected function setUp()
+    {
+        $this->deleteTestFiles();
+
+        $this->locator = new FileLocator();
+    }
+
+    public function testItFindsFilesByGlobbing()
+    {
+        $this->createTestFile('config1.php');
+        $this->createTestFile('config2.php');
+        $this->createTestFile('config.json');
+
+        $files = $this->locator->locate([$this->getTestPath('*.php')]);
+
+        $this->assertEquals([
+            $this->getTestPath('config1.php'),
+            $this->getTestPath('config2.php'),
+        ], $files);
+    }
+}